move tests to bellperson adapter and remove tests folder; make bellperson adapter private
